refactor(tooltip): distribute getClosestData logic into per-series files
